`typechecker` is a module now
