+++ ChoreographyEngine/utils/typechecker/README.md
@@ -1,0 +1,29 @@
+# `typechecker`
+
+![](https://img.shields.io/badge/Python-3.6-blue.svg) [![TEST - PASS](https://img.shields.io/badge/TEST-PASS-green.svg "Click to See Test Samples")](test.py)
+
+A type checker for Python data types.
+
+## Supported Types
+
+- Simple Data Types:
+  - Number: `int`, `float`, `complex`
+  - String: `str`
+  - Bool: `bool`
+  - Others: User-defined and so on...
+- Compound Data Types:
+  - Tuple: `tuple`
+  - Sets: `set`
+  - List: `list`
+  - Dictionary: `dict`
+
+## Some Restrictions on Usage
+
+1. Not really Python type but static data type;
+2. Only available for the 4 compound data;
+3. Cannot check arguments' type of functions.
+
+## How to Use 
+
+It is very easy. You can see samples in [test.py](test.py).
+
