+++ ChoreographyEngine/utils/typechecker/test.py
@@ -1,45 +1,7 @@
 #! python3
 # coding: utf-8
 
-"""
-A Type Checker for Python Basic Data Types:
-    Simple Data Types:
-        Number: int, float, complex
-        String: str
-        Bool: bool
-        Others: User-defined and so on...
-    Compound Data Types:
-        Tuple: tuple
-        Sets: set
-        List: list
-        Dictionary: dict
-* Some restriction of usage:
-    1. Not really Python type but static data type;
-    2. Only available for the 4 compound data;
-    3. Cannot check arguments' type of functions.
-"""
-
-function_type = type(lambda _: _)
-
-
-def check_type(element, some_type) -> bool:
-    compound_types = {tuple, set, list, dict}
-    element_type = type(element)
-    if element_type is type(some_type) and element_type in compound_types:
-        if element_type is tuple:
-            return len(element) == len(some_type) and \
-                all(map((lambda x: check_type(x[0], x[1])), zip(element, some_type)))
-        if element_type is list:
-            return all(map((lambda x: check_type(x, some_type[0])), element))
-        if element_type is set:
-            return all(map((lambda x: check_type(x, tuple(some_type)[0])), element))
-        if element_type is dict:
-            key_type = tuple(some_type)[0]
-            return all(map((lambda x: check_type(x, key_type)), (i for i in element))) and \
-                all(map((lambda x: check_type(x, some_type[key_type])), (element[i] for i in element)))
-        return False
-    return type(element) is some_type
-
+from typechecker import function_type, check_type
 
 # Simple tests:
 if __name__ == '__main__':
@@ -94,7 +56,7 @@ if __name__ == '__main__':
         ],
         'Test compound types as simple data types': [
             ({}, dict, True),
-            ({1: '', '': 2,  None: None}, dict, True),
+            ({1: '', '': 2, None: None}, dict, True),
             ([], list, True),
             ([1, ''], list, True),
             ((1, 2, 3), tuple, True),
@@ -121,6 +83,7 @@ if __name__ == '__main__':
             else:
                 print('FAILED')
                 import sys
+
                 sys.stderr.write("    EXPECTED: %s\n    BUT GOT: %s\n" %
                                  (str(expected_result), str(real_result)))
                 sys.exit(2)
